pyproject.toml
[tool.poetry]
name = "layout2svg"
version = "0.1.0"
description = ""
authors = ["dasdias <das.dias6@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
docopt = "^0.6.2"
klayout = "^0.28.10"
loguru = "^0.7.0"
gdstk = "^0.9.42"
PyYAML = "^6.0.1"
colour = "^0.1.5"
lxml = "^4.9.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^4.1.0"
black = "^23.7.0"

[tool.poetry.scripts]
layout2svg = "bin/layout2svg"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
